+++ src/utils/excel.ts
@@ -89,6 +89,7 @@ export const excelExport = (data: any) => {
     addSheet('BrouwdagStappen',       data.brouwdag_stappen)
     addSheet('WaterAddities',         data.water_addities)
     addSheet('WaterProfielen',        data.water_profielen)
+    addSheet('WaterDoelprofielen',    data.water_doelprofielen)
     addSheet('HopAddities',           data.hop_addities)
     addSheet('DryHops',               data.dry_hops)
     addSheet('KoelLogs',              data.koel_logs)
@@ -269,6 +270,7 @@ export const excelImport = (file: File, cb: (data: any) => void, onError?: () =>
         brouwdag_stappen:             parse('BrouwdagStappen'),
         water_addities:               parse('WaterAddities'),
         water_profielen:              parse('WaterProfielen'),
+        water_doelprofielen:          parse('WaterDoelprofielen'),
         hop_addities:                 parse('HopAddities'),
         dry_hops:                     parse('DryHops'),
         koel_logs:                    parse('KoelLogs'),